redundant code found in Primary Header, it should have been using Login
