+++ app-web/src/components/PrimaryHeader/PrimaryHeader.module.css
@@ -7,7 +7,3 @@
     align-items: center;
     justify-content: space-between
 }
-
-.PrimaryHeader button {
-    display: none;
-}
